requirements.txt
flask
googlesearch-python
requests
beautifulsoup4